case 6 breaks cuz of type mismatch, think i need do more of this in the actual files rather than in the runner

+++ FractionalDynamic.java
@@ -15,7 +15,7 @@ public class FractionalDynamic {
             }
         }
 
-        return dp[n][(int) capacity];
+        return dp[n][capacity];
     }
 }
